Move model to correct namespace
