+++ source/test-data-generation/GreenEnergyHub.Aggregation.TestData.Application/Parsers/ChargePriceTestDataParser.cs
@@ -15,6 +15,7 @@
 using GreenEnergyHub.Aggregation.TestData.Application.Service;
 using GreenEnergyHub.Aggregation.TestData.Infrastructure;
 using GreenEnergyHub.Aggregation.TestData.Infrastructure.CosmosDb;
+using GreenEnergyHub.Aggregation.TestData.Infrastructure.Models;
 
 namespace GreenEnergyHub.Aggregation.TestData.Application.Parsers
 {
